update to data fetching
